backend/requirements.txt
beautifulsoup4==4.12.2
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==41.0.7
feedparser==6.0.10
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.30.0
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
gunicorn==21.2.0
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
oauthlib==3.3.1
packaging==26.0
playwright==1.58.0
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyee==13.0.1
PyJWT==2.8.0
pyparsing==3.3.2
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
sgmllib3k==1.0.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
validators==0.22.0
Werkzeug==3.1.6
